Packages by Scott Grayban; tone down contributions section.
Scott Grayban is providing pre-built binaries for Debian.

The code contributions section was too aggressive in the phrasing used
for code submission authority; I knew that at the time, but went with
it.  Rephrased to remove the negative terms and confrontational attitude
and instead says what the submitter is implicitly doing by making a
submission.

[ci skip]