2 bugs: HTML non-template, & outbound hostgraph
A variable reference in template HTML was missing its {{...}} wrapper,
so was making it through to the emitted HTML as a malformed rowspan
value.  Fixed.

When walking the hostgraph entries, some corner case means that it's
possible to not have entries in the outbound hostgraph.  That's a bug,
but the walking code would panic when seeing this.  When running
normally, that panic does not make it to the logs, so we get unlogged
silent failures.

Instead, check for existence of the relevant nodes, and log
non-existence so we can tell that something hinky is happening.

This still leaves the code-path which leaves the outbound hostgraph
entry missing (but not the inbound, and not if reloading from persisted
JSON).  Hacky fast weekend project bites again.