Drop unused flag, add context for out-queries
The `-timeout-stats-fetch` flag was unused, but `-http-fetch-timeout`
was used; they were defaulting to 30s and 120s respectively.  Drop the
former, set the remaining default to the former's.

Use a context with deadline expiration around the HTTP request as a
whole, to cover more than just timeout fetching headers.  We should
already have DNS resolved before getting here, so probably not much to
it.